+++ src/transport/http/server/server.rs
@@ -25,7 +25,7 @@
 // DEALINGS IN THE SOFTWARE.
 
 use crate::common;
-use crate::http::server::background;
+use crate::transport::http::server::background;
 use crate::transport::{TransportServer, TransportServerEvent};
 use crate::server_utils::access_control::AccessControl;
